# Speed LTO builds by using multiple threads